+++ partiql-catalog/src/call_defs.rs
@@ -4,6 +4,7 @@ use partiql_logical::ValueExpr;
 use std::fmt::{Debug, Formatter};
 use thiserror::Error;
 
+use crate::scalar_fn::ScalarFnExpr;
 use unicase::UniCase;
 
 /// An error that can happen during call lookup
@@ -15,13 +16,6 @@ pub enum CallLookupError {
     InvalidNumberOfArguments(String),
 }
 
-#[derive(Debug, Eq, PartialEq)]
-pub enum CallArgument {
-    Positional(ValueExpr),
-    Named(String, ValueExpr),
-    Star,
-}
-
 #[derive(Debug)]
 pub struct CallDef {
     pub names: Vec<&'static str>,
@@ -29,6 +23,7 @@ pub struct CallDef {
 }
 
 impl CallDef {
+    // Used when lowering AST -> plan
     pub fn lookup(&self, args: &[CallArgument], name: &str) -> Result<ValueExpr, CallLookupError> {
         'overload: for overload in &self.overloads {
             let formals = &overload.input;
@@ -53,6 +48,13 @@ impl CallDef {
     }
 }
 
+#[derive(Debug, Eq, PartialEq)]
+pub enum CallArgument {
+    Positional(ValueExpr),
+    Named(String, ValueExpr),
+    Star,
+}
+
 #[derive(Debug, Copy, Clone)]
 pub enum CallSpecArg {
     Positional,
@@ -75,8 +77,6 @@ impl CallSpecArg {
     }
 }
 
-impl CallSpecArg {}
-
 pub struct CallSpec {
     pub input: Vec<CallSpecArg>,
     pub output: Box<dyn Fn(Vec<ValueExpr>) -> logical::ValueExpr + Send + Sync>,
@@ -85,5 +85,26 @@ pub struct CallSpec {
 impl Debug for CallSpec {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "CallSpec [{:?}]", &self.input)
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct ScalarFnCallDef {
+    pub names: Vec<&'static str>,
+    pub overloads: ScalarFnCallSpecs,
+}
+
+pub type ScalarFnCallSpecs = Vec<ScalarFnCallSpec>;
+
+#[derive(Clone)]
+pub struct ScalarFnCallSpec {
+    // TODO:  Include Scalar Function attributes (e.g., isNullCall and isMissingCall, etc.): https://github.com/partiql/partiql-lang-rust/issues/499
+    pub input: Vec<CallSpecArg>,
+    pub output: Box<dyn ScalarFnExpr>,
+}
+
+impl Debug for ScalarFnCallSpec {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "ScalarFnCallSpec [{:?}]", &self.input)
     }
 }