+++ partiql/benches/bench_eval_multi_like.rs
@@ -2,15 +2,14 @@ use std::time::Duration;
 
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use itertools::Itertools;
+use partiql_catalog::catalog::{Catalog, PartiqlCatalog};
 use partiql_catalog::context::SystemContext;
-use partiql_catalog::{Catalog, PartiqlCatalog};
-use rand::{Rng, SeedableRng};
-
 use partiql_eval::env::basic::MapBindings;
 use partiql_eval::eval::{BasicContext, EvalPlan};
 use partiql_eval::plan::{EvaluationMode, EvaluatorPlanner};
 use partiql_logical::{BindingsOp, LogicalPlan};
 use partiql_logical_planner::LogicalPlanner;
+use rand::{Rng, SeedableRng};
 
 use partiql_parser::{Parser, ParserResult};
 use partiql_value::{tuple, Bag, DateTime, Value};