+++ partiql/tests/extension_error.rs
@@ -6,10 +6,11 @@ use std::error::Error;
 use thiserror::Error;
 
 use partiql_catalog::call_defs::{CallDef, CallSpec, CallSpecArg};
+use partiql_catalog::catalog::{Catalog, PartiqlCatalog};
 use partiql_catalog::context::{SessionContext, SystemContext};
-use partiql_catalog::{
-    BaseTableExpr, BaseTableExprResult, BaseTableExprResultError, BaseTableFunctionInfo, Catalog,
-    Extension, PartiqlCatalog, TableFunction,
+use partiql_catalog::extension::{Extension, ExtensionResultError};
+use partiql_catalog::table_fn::{
+    BaseTableExpr, BaseTableExprResult, BaseTableFunctionInfo, TableFunction,
 };
 use partiql_eval::env::basic::MapBindings;
 use partiql_eval::error::{EvalErr, EvaluationError};
@@ -23,7 +24,7 @@ use partiql_logical as logical;
 #[derive(Debug)]
 pub struct UserCtxTestExtension {}
 
-impl partiql_catalog::Extension for UserCtxTestExtension {
+impl partiql_catalog::extension::Extension for UserCtxTestExtension {
     fn name(&self) -> String {
         "test_extension".into()
     }
@@ -95,12 +96,12 @@ impl BaseTableExpr for EvalTestCtxTable {
                 Value::String(_name) => Ok(Box::new(TestDataGen {})),
                 _ => {
                     let error = UserCtxError::BadArgs;
-                    Err(Box::new(error) as BaseTableExprResultError)
+                    Err(Box::new(error) as ExtensionResultError)
                 }
             }
         } else {
             let error = UserCtxError::BadArgs;
-            Err(Box::new(error) as BaseTableExprResultError)
+            Err(Box::new(error) as ExtensionResultError)
         }
     }
 }
@@ -108,7 +109,7 @@ impl BaseTableExpr for EvalTestCtxTable {
 struct TestDataGen {}
 
 impl Iterator for TestDataGen {
-    type Item = Result<Value, BaseTableExprResultError>;
+    type Item = Result<Value, ExtensionResultError>;
 
     fn next(&mut self) -> Option<Self::Item> {
         Some(Err(Box::new(UserCtxError::Runtime)))